+++ scripts/build-styles.mjs
@@ -38,6 +38,7 @@ await writeFile(
     '@import "../../src/components/button/button.css";',
     '@import "../../src/components/icon-button/icon-button.css";',
     '@import "../../src/components/icon/icon.css";',
+    '@import "../../src/components/image/image.css";',
     '@import "../../src/components/app-bar/app-bar.css";',
     '@import "../../src/components/card/card.css";',
     '@import "../../src/components/dialog/dialog.css";',