+++ src/image.ts
@@ -1,0 +1,12 @@
+"use client";
+
+export {
+  Image,
+  type ImageContentProps,
+  type ImageFallbackProps,
+  type ImageFit,
+  type ImageFrame,
+  type ImagePosition,
+  type ImageRadius,
+  type ImageRootProps,
+} from "./components/image/index.js";